setup new routes within graph for weekly and monthly views
