--- frontend/src/components/Graph.js
@@ -1,84 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { getEntries } from '../actions';
-import Swipeable from 'react-swipeable';
-import demoGraph from '../assets/graph.png';
-
-class Graph extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      entries: []
-    };
-  }
-
-  handleSwipeLeft() {
-    alert('Swiped left!');
-  }
-
-  handleSwipeRight() {
-    alert('Swiped right!');
-  }
-
-  componentWillMount() {
-    this.props.getEntries();
-  }
-
-  render() {
-    console.log(this.props.entries);
-    return (
-      <Swipeable
-        onSwipedLeft={this.handleSwipeLeft}
-        onSwipedRight={this.handleSwipeRight}
-      >
-        <div className="container is-mobile" id="mainBox">
-          <img src={demoGraph} alt="demo graph" />
-          <div className="columns is-mobile">
-            <div className="column">
-              <button className="button is-danger">Current Week</button>
-            </div>
-            <div className="column">
-              <button className="button is-danger">Month</button>
-            </div>
-          </div>
-          {this.props.entries.map(entry => {
-            return (
-              <article key={entry.id} className="media">
-                <figure className="media-left">
-                  <i className="fa fa-heart" />
-                </figure>
-                <div className="media-content">
-                  <div className="content" id="entryText">
-                    <p>
-                      <strong>Entry {entry.id} </strong>
-                      <small>{entry.createdAt}</small>
-                      <br />
-                      {entry.text}
-                    </p>
-                  </div>
-                </div>
-              </article>
-            );
-          })}
-        </div>
-      </Swipeable>
-    );
-  }
-}
-
-const mapStatetoProps = state => {
-  return { entries: state.entries };
-};
-
-const mapDispatchtoProps = dispatch => {
-  return {
-    getEntries: () => {
-      dispatch(getEntries());
-    }
-  };
-};
-
-const ConnectedGraph = connect(mapStatetoProps, mapDispatchtoProps)(Graph);
-
-export default ConnectedGraph;
